@@ -2,7 +2,7 @@ package common
 
 const (
 	SERVICENAME = "casaos"
-	VERSION     = "0.4.4.3"
+	VERSION     = "0.4.15"
 	BODY        = " "
 	RANW_NAME   = "IceWhale-RemoteAccess"
 )
