@@ -11,33 +11,14 @@ import (
 	"github.com/IceWhaleTech/CasaOS/pkg/utils/common_err"
 	"github.com/IceWhaleTech/CasaOS/pkg/utils/httper"
 	"github.com/IceWhaleTech/CasaOS/service"
-	"github.com/gin-gonic/gin"
+	"github.com/labstack/echo/v4"
 	"go.uber.org/zap"
 )
 
-func ListStorages(c *gin.Context) {
-	// var req model.PageReq
-	// if err := c.ShouldBind(&req); err != nil {
-	// 	c.JSON(common_err.SUCCESS, model.Result{Success: common_err.CLIENT_ERROR, Message: common_err.GetMsg(common_err.CLIENT_ERROR), Data: err.Error()})
-	// 	return
-	// }
-	// req.Validate()
-
-	//logger.Info("ListStorages", zap.Any("req", req))
-	//storages, total, err := service.MyService.Storage().GetStorages(req.Page, req.PerPage)
-	// if err != nil {
-	// 	c.JSON(common_err.SUCCESS, model.Result{Success: common_err.SERVICE_ERROR, Message: common_err.GetMsg(common_err.SERVICE_ERROR), Data: err.Error()})
-	// 	return
-	// }
-	// c.JSON(common_err.SUCCESS, model.Result{Success: common_err.SUCCESS, Message: common_err.GetMsg(common_err.SUCCESS), Data: model.PageResp{
-	// 	Content: storages,
-	// 	Total:   total,
-	// }})
+func ListStorages(ctx echo.Context) error {
 	r, err := service.MyService.Storage().GetStorages()
-
 	if err != nil {
-		c.JSON(common_err.SUCCESS, model.Result{Success: common_err.SERVICE_ERROR, Message: common_err.GetMsg(common_err.SERVICE_ERROR), Data: err.Error()})
-		return
+		return ctx.JSON(common_err.SUCCESS, model.Result{Success: common_err.SERVICE_ERROR, Message: common_err.GetMsg(common_err.SERVICE_ERROR), Data: err.Error()})
 	}
 
 	for i := 0; i < len(r.MountPoints); i++ {
@@ -53,7 +34,6 @@ func ListStorages(c *gin.Context) {
 			r.MountPoints[i].Icon = dropbox.ICONURL
 		}
 		if dataMap["type"] == "onedrive" {
-
 			r.MountPoints[i].Icon = onedrive.ICONURL
 		}
 		r.MountPoints[i].Name = dataMap["username"]
@@ -61,46 +41,39 @@ func ListStorages(c *gin.Context) {
 	list := []httper.MountPoint{}
 
 	for _, v := range r.MountPoints {
-		list = append(list, httper.MountPoint{
-			Fs:         v.Fs,
-			Icon:       v.Icon,
-			MountPoint: v.MountPoint,
-			Name:       v.Name,
-		})
+		list = append(list, httper.MountPoint(v))
 	}
 
-	c.JSON(common_err.SUCCESS, model.Result{Success: common_err.SUCCESS, Message: common_err.GetMsg(common_err.SUCCESS), Data: list})
+	return ctx.JSON(common_err.SUCCESS, model.Result{Success: common_err.SUCCESS, Message: common_err.GetMsg(common_err.SUCCESS), Data: list})
 }
 
-func UmountStorage(c *gin.Context) {
+func UmountStorage(ctx echo.Context) error {
 	json := make(map[string]string)
-	c.ShouldBind(&json)
+	ctx.Bind(&json)
 	mountPoint := json["mount_point"]
 	if mountPoint == "" {
-		c.JSON(common_err.CLIENT_ERROR, model.Result{Success: common_err.CLIENT_ERROR, Message: common_err.GetMsg(common_err.CLIENT_ERROR), Data: "mount_point is empty"})
-		return
+		return ctx.JSON(common_err.CLIENT_ERROR, model.Result{Success: common_err.CLIENT_ERROR, Message: common_err.GetMsg(common_err.CLIENT_ERROR), Data: "mount_point is empty"})
 	}
 	err := service.MyService.Storage().UnmountStorage(mountPoint)
 	if err != nil {
-		c.JSON(common_err.SERVICE_ERROR, model.Result{Success: common_err.SERVICE_ERROR, Message: common_err.GetMsg(common_err.SERVICE_ERROR), Data: err.Error()})
-		return
+		return ctx.JSON(common_err.SERVICE_ERROR, model.Result{Success: common_err.SERVICE_ERROR, Message: common_err.GetMsg(common_err.SERVICE_ERROR), Data: err.Error()})
 	}
 	service.MyService.Storage().DeleteConfigByName(strings.ReplaceAll(mountPoint, "/mnt/", ""))
-	c.JSON(common_err.SUCCESS, model.Result{Success: common_err.SUCCESS, Message: common_err.GetMsg(common_err.SUCCESS), Data: "success"})
+	return ctx.JSON(common_err.SUCCESS, model.Result{Success: common_err.SUCCESS, Message: common_err.GetMsg(common_err.SUCCESS), Data: "success"})
 }
 
-func GetStorage(c *gin.Context) {
-
-	// idStr := c.Query("id")
+func GetStorage(ctx echo.Context) error {
+	// idStr := ctx.QueryParam("id")
 	// id, err := strconv.Atoi(idStr)
 	// if err != nil {
-	// 	c.JSON(common_err.SUCCESS, model.Result{Success: common_err.CLIENT_ERROR, Message: common_err.GetMsg(common_err.CLIENT_ERROR), Data: err.Error()})
+	// 	return ctx.JSON(common_err.SUCCESS, model.Result{Success: common_err.CLIENT_ERROR, Message: common_err.GetMsg(common_err.CLIENT_ERROR), Data: err.Error()})
 	// 	return
 	// }
 	// storage, err := service.MyService.Storage().GetStorageById(uint(id))
 	// if err != nil {
-	// 	c.JSON(common_err.SUCCESS, model.Result{Success: common_err.SERVICE_ERROR, Message: common_err.GetMsg(common_err.SERVICE_ERROR), Data: err.Error()})
+	// 	return ctx.JSON(common_err.SUCCESS, model.Result{Success: common_err.SERVICE_ERROR, Message: common_err.GetMsg(common_err.SERVICE_ERROR), Data: err.Error()})
 	// 	return
 	// }
-	// c.JSON(common_err.SUCCESS, model.Result{Success: common_err.SUCCESS, Message: common_err.GetMsg(common_err.SUCCESS), Data: storage})
+	// return ctx.JSON(common_err.SUCCESS, model.Result{Success: common_err.SUCCESS, Message: common_err.GetMsg(common_err.SUCCESS), Data: storage})
+	return nil
 }
