@@ -6,8 +6,8 @@ import (
 
 	"github.com/IceWhaleTech/CasaOS/service"
 	"github.com/IceWhaleTech/CasaOS/types"
-	"github.com/gin-gonic/gin"
 	"github.com/gorilla/websocket"
+	"github.com/labstack/echo/v4"
 )
 
 var upGrader = websocket.Upgrader{
@@ -24,11 +24,11 @@ var upGrader = websocket.Upgrader{
 // @Param token path string true "token"
 // @Success 200 {string} string "ok"
 // @Router /notify/ws [get]
-func NotifyWS(c *gin.Context) {
-	//升级get请求为webSocket协议
-	ws, err := upGrader.Upgrade(c.Writer, c.Request, nil)
+func NotifyWS(ctx echo.Context) error {
+	// 升级get请求为webSocket协议
+	ws, err := upGrader.Upgrade(ctx.Response().Writer, ctx.Request(), nil)
 	if err != nil {
-		return
+		return nil
 	}
 	defer ws.Close()
 	service.WebSocketConns = append(service.WebSocketConns, ws)
@@ -41,7 +41,6 @@ func NotifyWS(c *gin.Context) {
 		mt, message, err := ws.ReadMessage()
 		fmt.Println(mt, message, err)
 	}
-
 }
 
 // @Summary 标记notify已读
@@ -51,12 +50,13 @@ func NotifyWS(c *gin.Context) {
 // @Security ApiKeyAuth
 // @Success 200 {string} string "ok"
 // @Router /notify/read/{id} [put]
-func PutNotifyRead(c *gin.Context) {
-	id := c.Param("id")
+func PutNotifyRead(ctx echo.Context) error {
+	id := ctx.Param("id")
 	// if len(id) == 0 {
-	// 	c.JSON(http.StatusOK, model.Result{Success: oasis_err.INVALID_PARAMS, Message: oasis_err.GetMsg(oasis_err.INVALID_PARAMS)})
+	// 	return ctx.JSON(http.StatusOK, model.Result{Success: oasis_err.INVALID_PARAMS, Message: oasis_err.GetMsg(oasis_err.INVALID_PARAMS)})
 	// 	return
 	// }
 	fmt.Println(id)
 	service.MyService.Notify().MarkRead(id, types.NOTIFY_READ)
+	return nil
 }
