@@ -1,6 +1,8 @@
 package v1
 
 import (
+	"context"
+	"net/http"
 	"strconv"
 	"strings"
 	"time"
@@ -10,57 +12,52 @@ import (
 	"github.com/IceWhaleTech/CasaOS/drivers/google_drive"
 	"github.com/IceWhaleTech/CasaOS/drivers/onedrive"
 	"github.com/IceWhaleTech/CasaOS/service"
-	"github.com/gin-gonic/gin"
+	"github.com/labstack/echo/v4"
 	"go.uber.org/zap"
 )
 
-func GetRecoverStorage(c *gin.Context) {
-	c.Header("Content-Type", "text/html; charset=utf-8")
-	t := c.Param("type")
+func GetRecoverStorage(ctx echo.Context) error {
+	t := strings.TrimSuffix(ctx.Param("type"), "/")
 	currentTime := time.Now().UTC()
 	currentDate := time.Now().UTC().Format("2006-01-02")
 	notify := make(map[string]interface{})
+	event := "casaos:file:recover"
 	if t == "GoogleDrive" {
 		google_drive := google_drive.GetConfig()
-		google_drive.Code = c.Query("code")
+		google_drive.Code = ctx.QueryParam("code")
 		if len(google_drive.Code) == 0 {
-			c.String(200, `<p>Code cannot be empty</p><script>window.close()</script>`)
 			notify["status"] = "fail"
 			notify["message"] = "Code cannot be empty"
 			logger.Error("Then code is empty: ", zap.String("code", google_drive.Code), zap.Any("name", "google_drive"))
 			service.MyService.Notify().SendNotify("casaos:file:recover", notify)
-			return
+			return ctx.HTML(http.StatusOK, `<p>Code cannot be empty</p><script>window.close()</script>`)
 		}
-
-		err := google_drive.Init(c)
+		err := google_drive.Init(context.Background())
 		if err != nil {
-			c.String(200, `<p>Initialization failure:`+err.Error()+`</p><script>window.close()</script>`)
 			notify["status"] = "fail"
 			notify["message"] = "Initialization failure"
 			logger.Error("Then init error: ", zap.Error(err), zap.Any("name", "google_drive"))
-			service.MyService.Notify().SendNotify("casaos:file:recover", notify)
-			return
+			service.MyService.Notify().SendNotify(event, notify)
+			return ctx.HTML(http.StatusOK, `<p>Initialization failure:`+err.Error()+`</p><script>window.close()</script>`)
 		}
 
-		username, err := google_drive.GetUserInfo(c)
+		username, err := google_drive.GetUserInfo(context.Background())
 		if err != nil {
-			c.String(200, `<p>Failed to get user information:`+err.Error()+`</p><script>window.close()</script>`)
 			notify["status"] = "fail"
 			notify["message"] = "Failed to get user information"
 			logger.Error("Then get user info error: ", zap.Error(err), zap.Any("name", "google_drive"))
-			service.MyService.Notify().SendNotify("casaos:file:recover", notify)
-			return
+			service.MyService.Notify().SendNotify(event, notify)
+			return ctx.HTML(http.StatusOK, `<p>Failed to get user information:`+err.Error()+`</p><script>window.close()</script>`)
 		}
 		dmap := make(map[string]string)
 		dmap["username"] = username
 		configs, err := service.MyService.Storage().GetConfig()
 		if err != nil {
-			c.String(200, `<p>Failed to get rclone config:`+err.Error()+`</p><script>window.close()</script>`)
 			notify["status"] = "fail"
 			notify["message"] = "Failed to get rclone config"
 			logger.Error("Then get config error: ", zap.Error(err), zap.Any("name", "google_drive"))
-			service.MyService.Notify().SendNotify("casaos:file:recover", notify)
-			return
+			service.MyService.Notify().SendNotify(event, notify)
+			return ctx.HTML(http.StatusOK, `<p>Failed to get rclone config:`+err.Error()+`</p><script>window.close()</script>`)
 		}
 		for _, v := range configs.Remotes {
 			cf, err := service.MyService.Storage().GetConfigByName(v)
@@ -69,15 +66,14 @@ func GetRecoverStorage(c *gin.Context) {
 				continue
 			}
 			if cf["type"] == "drive" && cf["username"] == dmap["username"] {
-				c.String(200, `<p>The same configuration has been added</p><script>window.close()</script>`)
 				err := service.MyService.Storage().CheckAndMountByName(v)
 				if err != nil {
 					logger.Error("check and mount by name error: ", zap.Error(err), zap.Any("name", cf["username"]))
 				}
 				notify["status"] = "warn"
 				notify["message"] = "The same configuration has been added"
-				service.MyService.Notify().SendNotify("casaos:file:recover", notify)
-				return
+				service.MyService.Notify().SendNotify(event, notify)
+				return ctx.HTML(http.StatusOK, `<p>The same configuration has been added</p><script>window.close()</script>`)
 			}
 		}
 		if len(username) > 0 {
@@ -85,7 +81,6 @@ func GetRecoverStorage(c *gin.Context) {
 			username = a[0]
 		}
 
-		//username = fileutil.NameAccumulation(username, "/mnt")
 		username += "_google_drive_" + strconv.FormatInt(time.Now().Unix(), 10)
 
 		dmap["client_id"] = google_drive.ClientID
@@ -99,48 +94,45 @@ func GetRecoverStorage(c *gin.Context) {
 		notify["status"] = "success"
 		notify["message"] = "Success"
 		notify["driver"] = "GoogleDrive"
-		service.MyService.Notify().SendNotify("casaos:file:recover", notify)
+		service.MyService.Notify().SendNotify(event, notify)
 	} else if t == "Dropbox" {
 		dropbox := dropbox.GetConfig()
-		dropbox.Code = c.Query("code")
+
+		dropbox.Code = ctx.QueryParam("code")
 		if len(dropbox.Code) == 0 {
-			c.String(200, `<p>Code cannot be empty</p><script>window.close()</script>`)
+
 			notify["status"] = "fail"
 			notify["message"] = "Code cannot be empty"
 			logger.Error("Then code is empty error: ", zap.String("code", dropbox.Code), zap.Any("name", "dropbox"))
-			service.MyService.Notify().SendNotify("casaos:file:recover", notify)
-			return
+			service.MyService.Notify().SendNotify(event, notify)
+			return ctx.HTML(http.StatusOK, `<p>Code cannot be empty</p><script>window.close()</script>`)
 		}
-
-		err := dropbox.Init(c)
+		err := dropbox.Init(context.Background())
 		if err != nil {
-			c.String(200, `<p>Initialization failure:`+err.Error()+`</p><script>window.close()</script>`)
 			notify["status"] = "fail"
 			notify["message"] = "Initialization failure"
 			logger.Error("Then init error: ", zap.Error(err), zap.Any("name", "dropbox"))
-			service.MyService.Notify().SendNotify("casaos:file:recover", notify)
-			return
+			service.MyService.Notify().SendNotify(event, notify)
+			return ctx.HTML(http.StatusOK, `<p>Initialization failure:`+err.Error()+`</p><script>window.close()</script>`)
 		}
-		username, err := dropbox.GetUserInfo(c)
+		username, err := dropbox.GetUserInfo(context.Background())
 		if err != nil {
-			c.String(200, `<p>Failed to get user information:`+err.Error()+`</p><script>window.close()</script>`)
 			notify["status"] = "fail"
 			notify["message"] = "Failed to get user information"
 			logger.Error("Then get user information: ", zap.Error(err), zap.Any("name", "dropbox"))
-			service.MyService.Notify().SendNotify("casaos:file:recover", notify)
-			return
+			service.MyService.Notify().SendNotify(event, notify)
+			return ctx.HTML(http.StatusOK, `<p>Failed to get user information:`+err.Error()+`</p><script>window.close()</script>`)
 		}
 		dmap := make(map[string]string)
 		dmap["username"] = username
 
 		configs, err := service.MyService.Storage().GetConfig()
 		if err != nil {
-			c.String(200, `<p>Failed to get rclone config:`+err.Error()+`</p><script>window.close()</script>`)
 			notify["status"] = "fail"
 			notify["message"] = "Failed to get rclone config"
 			logger.Error("Then get config error: ", zap.Error(err), zap.Any("name", "dropbox"))
-			service.MyService.Notify().SendNotify("casaos:file:recover", notify)
-			return
+			service.MyService.Notify().SendNotify(event, notify)
+			return ctx.HTML(http.StatusOK, `<p>Failed to get rclone config:`+err.Error()+`</p><script>window.close()</script>`)
 		}
 		for _, v := range configs.Remotes {
 			cf, err := service.MyService.Storage().GetConfigByName(v)
@@ -149,16 +141,14 @@ func GetRecoverStorage(c *gin.Context) {
 				continue
 			}
 			if cf["type"] == "dropbox" && cf["username"] == dmap["username"] {
-				c.String(200, `<p>The same configuration has been added</p><script>window.close()</script>`)
-				err := service.MyService.Storage().CheckAndMountByName(v)
-				if err != nil {
+				if err := service.MyService.Storage().CheckAndMountByName(v); err != nil {
 					logger.Error("check and mount by name error: ", zap.Error(err), zap.Any("name", cf["username"]))
 				}
 
 				notify["status"] = "warn"
 				notify["message"] = "The same configuration has been added"
-				service.MyService.Notify().SendNotify("casaos:file:recover", notify)
-				return
+				service.MyService.Notify().SendNotify(event, notify)
+				return ctx.HTML(http.StatusOK, `<p>The same configuration has been added</p><script>window.close()</script>`)
 			}
 		}
 		if len(username) > 0 {
@@ -171,66 +161,49 @@ func GetRecoverStorage(c *gin.Context) {
 		dmap["client_secret"] = dropbox.AppSecret
 		dmap["token"] = `{"access_token":"` + dropbox.AccessToken + `","token_type":"bearer","refresh_token":"` + dropbox.Addition.RefreshToken + `","expiry":"` + currentDate + `T` + currentTime.Add(time.Hour*3).Add(time.Minute*50).Format("15:04:05") + `.780385354Z"}`
 		dmap["mount_point"] = "/mnt/" + username
-		// data.SetValue(username, "type", "dropbox")
-		// data.SetValue(username, "client_id", add.AppKey)
-		// data.SetValue(username, "client_secret", add.AppSecret)
-		// data.SetValue(username, "mount_point", "/mnt/"+username)
-
-		// data.SetValue(username, "token", `{"access_token":"`+dropbox.AccessToken+`","token_type":"bearer","refresh_token":"`+dropbox.Addition.RefreshToken+`","expiry":"`+currentDate+`T`+currentTime.Add(time.Hour*3).Format("15:04:05")+`.780385354Z"}`)
-		// e = data.Save()
-		// if e != nil {
-		// 	c.String(200, `<p>保存配置失败:`+e.Error()+`</p>`)
-
-		// 	return
-		// }
 		service.MyService.Storage().CreateConfig(dmap, username, "dropbox")
 		service.MyService.Storage().MountStorage("/mnt/"+username, username+":")
 
 		notify["status"] = "success"
 		notify["message"] = "Success"
 		notify["driver"] = "Dropbox"
-		service.MyService.Notify().SendNotify("casaos:file:recover", notify)
+		service.MyService.Notify().SendNotify(event, notify)
 	} else if t == "Onedrive" {
 		onedrive := onedrive.GetConfig()
-		onedrive.Code = c.Query("code")
+		onedrive.Code = ctx.QueryParam("code")
 		if len(onedrive.Code) == 0 {
-			c.String(200, `<p>Code cannot be empty</p><script>window.close()</script>`)
 			notify["status"] = "fail"
 			notify["message"] = "Code cannot be empty"
 			logger.Error("Then code is empty error: ", zap.String("code", onedrive.Code), zap.Any("name", "onedrive"))
-			service.MyService.Notify().SendNotify("casaos:file:recover", notify)
-			return
+			service.MyService.Notify().SendNotify(event, notify)
+			return ctx.HTML(http.StatusOK, `<p>Code cannot be empty</p><script>window.close()</script>`)
 		}
-
-		err := onedrive.Init(c)
+		err := onedrive.Init(context.Background())
 		if err != nil {
-			c.String(200, `<p>Initialization failure:`+err.Error()+`</p><script>window.close()</script>`)
 			notify["status"] = "fail"
 			notify["message"] = "Initialization failure"
 			logger.Error("Then init error: ", zap.Error(err), zap.Any("name", "onedrive"))
-			service.MyService.Notify().SendNotify("casaos:file:recover", notify)
-			return
+			service.MyService.Notify().SendNotify(event, notify)
+			return ctx.HTML(http.StatusOK, `<p>Initialization failure:`+err.Error()+`</p><script>window.close()</script>`)
 		}
-		username, driveId, driveType, err := onedrive.GetInfo(c)
+		username, driveId, driveType, err := onedrive.GetInfo(context.Background())
 		if err != nil {
-			c.String(200, `<p>Failed to get user information:`+err.Error()+`</p><script>window.close()</script>`)
 			notify["status"] = "fail"
 			notify["message"] = "Failed to get user information"
 			logger.Error("Then get user information: ", zap.Error(err), zap.Any("name", "onedrive"))
-			service.MyService.Notify().SendNotify("casaos:file:recover", notify)
-			return
+			service.MyService.Notify().SendNotify(event, notify)
+			return ctx.HTML(http.StatusOK, `<p>Failed to get user information:`+err.Error()+`</p><script>window.close()</script>`)
 		}
 		dmap := make(map[string]string)
 		dmap["username"] = username
 
 		configs, err := service.MyService.Storage().GetConfig()
 		if err != nil {
-			c.String(200, `<p>Failed to get rclone config:`+err.Error()+`</p><script>window.close()</script>`)
 			notify["status"] = "fail"
 			notify["message"] = "Failed to get rclone config"
 			logger.Error("Then get config error: ", zap.Error(err), zap.Any("name", "onedrive"))
-			service.MyService.Notify().SendNotify("casaos:file:recover", notify)
-			return
+			service.MyService.Notify().SendNotify(event, notify)
+			return ctx.HTML(http.StatusOK, `<p>Failed to get rclone config:`+err.Error()+`</p><script>window.close()</script>`)
 		}
 		for _, v := range configs.Remotes {
 			cf, err := service.MyService.Storage().GetConfigByName(v)
@@ -239,16 +212,14 @@ func GetRecoverStorage(c *gin.Context) {
 				continue
 			}
 			if cf["type"] == "onedrive" && cf["username"] == dmap["username"] {
-				c.String(200, `<p>The same configuration has been added</p><script>window.close()</script>`)
-				err := service.MyService.Storage().CheckAndMountByName(v)
-				if err != nil {
+				if err := service.MyService.Storage().CheckAndMountByName(v); err != nil {
 					logger.Error("check and mount by name error: ", zap.Error(err), zap.Any("name", cf["username"]))
 				}
 
 				notify["status"] = "warn"
 				notify["message"] = "The same configuration has been added"
-				service.MyService.Notify().SendNotify("casaos:file:recover", notify)
-				return
+				service.MyService.Notify().SendNotify(event, notify)
+				return ctx.HTML(http.StatusOK, `<p>The same configuration has been added</p><script>window.close()</script>`)
 			}
 		}
 		if len(username) > 0 {
@@ -263,26 +234,14 @@ func GetRecoverStorage(c *gin.Context) {
 		dmap["mount_point"] = "/mnt/" + username
 		dmap["drive_id"] = driveId
 		dmap["drive_type"] = driveType
-		// data.SetValue(username, "type", "dropbox")
-		// data.SetValue(username, "client_id", add.AppKey)
-		// data.SetValue(username, "client_secret", add.AppSecret)
-		// data.SetValue(username, "mount_point", "/mnt/"+username)
-
-		// data.SetValue(username, "token", `{"access_token":"`+dropbox.AccessToken+`","token_type":"bearer","refresh_token":"`+dropbox.Addition.RefreshToken+`","expiry":"`+currentDate+`T`+currentTime.Add(time.Hour*3).Format("15:04:05")+`.780385354Z"}`)
-		// e = data.Save()
-		// if e != nil {
-		// 	c.String(200, `<p>保存配置失败:`+e.Error()+`</p>`)
-
-		// 	return
-		// }
 		service.MyService.Storage().CreateConfig(dmap, username, "onedrive")
 		service.MyService.Storage().MountStorage("/mnt/"+username, username+":")
 
 		notify["status"] = "success"
 		notify["message"] = "Success"
 		notify["driver"] = "Onedrive"
-		service.MyService.Notify().SendNotify("casaos:file:recover", notify)
+		service.MyService.Notify().SendNotify(event, notify)
 	}
 
-	c.String(200, `<p>Just close the page</p><script>window.close()</script>`)
+	return ctx.HTML(200, `<p>Just close the page</p><script>window.close()</script>`)
 }
