@@ -71,8 +71,7 @@ func InitV2Router() http.Handler {
 	e.Use(echo_middleware.JWTWithConfig(echo_middleware.JWTConfig{
 		Skipper: func(c echo.Context) bool {
 			return c.RealIP() == "::1" || c.RealIP() == "127.0.0.1"
-			//return true
-
+			// return true
 		},
 		ParseTokenFunc: func(token string, c echo.Context) (interface{}, error) {
 			valid, claims, err := jwt.Validate(token, func() (*ecdsa.PublicKey, error) { return external.GetPublicKey(config.CommonInfo.RuntimePath) })
@@ -84,11 +83,11 @@ func InitV2Router() http.Handler {
 			return claims, nil
 		},
 		TokenLookupFuncs: []echo_middleware.ValuesExtractor{
-			func(c echo.Context) ([]string, error) {
-				if len(c.Request().Header.Get(echo.HeaderAuthorization)) > 0 {
-					return []string{c.Request().Header.Get(echo.HeaderAuthorization)}, nil
+			func(ctx echo.Context) ([]string, error) {
+				if len(ctx.Request().Header.Get(echo.HeaderAuthorization)) > 0 {
+					return []string{ctx.Request().Header.Get(echo.HeaderAuthorization)}, nil
 				}
-				return []string{c.QueryParam("token")}, nil
+				return []string{ctx.QueryParam("token")}, nil
 			},
 		},
 	}))
@@ -116,6 +115,12 @@ func InitV2Router() http.Handler {
 	// })
 
 	e.Use(middleware.OapiRequestValidatorWithOptions(_swagger, &middleware.Options{
+		Skipper: func(c echo.Context) bool {
+			// jump validate when upload file
+			// because file upload can't pass validate
+			// issue: https://github.com/deepmap/oapi-codegen/issues/514
+			return strings.Contains(c.Request().Header[echo.HeaderContentType][0], "multipart/form-data")
+		},
 		Options: openapi3filter.Options{AuthenticationFunc: openapi3filter.NoopAuthenticationFunc},
 	}))
 
@@ -162,7 +167,7 @@ func InitFile() http.Handler {
 		fileName := path.Base(filePath)
 		w.Header().Add("Content-Disposition", "attachment; filename*=utf-8''"+url.PathEscape(fileName))
 		http.ServeFile(w, r, filePath)
-		//http.ServeFile(w, r, filePath)
+		// http.ServeFile(w, r, filePath)
 	})
 }
 
@@ -196,7 +201,7 @@ func InitDir() http.Handler {
 		list := strings.Split(files, ",")
 		for _, v := range list {
 			if !file.Exists(v) {
-				// c.JSON(common_err.SERVICE_ERROR, model.Result{
+				// return ctx.JSON(common_err.SERVICE_ERROR, model.Result{
 				// 	Success: common_err.FILE_DOES_NOT_EXIST,
 				// 	Message: common_err.GetMsg(common_err.FILE_DOES_NOT_EXIST),
 				// })
@@ -232,7 +237,7 @@ func InitDir() http.Handler {
 
 		err = ar.Create(w)
 		if err != nil {
-			//  c.JSON(common_err.SERVICE_ERROR, model.Result{
+			//  return ctx.JSON(common_err.SERVICE_ERROR, model.Result{
 			// 	Success: common_err.SERVICE_ERROR,
 			// 	Message: common_err.GetMsg(common_err.SERVICE_ERROR),
 			// 	Data:    err.Error(),
