@@ -2,10 +2,15 @@ package v2
 
 import (
 	"github.com/IceWhaleTech/CasaOS/codegen"
+	"github.com/IceWhaleTech/CasaOS/service"
 )
 
-type CasaOS struct{}
+type CasaOS struct {
+	fileUploadService *service.FileUploadService
+}
 
 func NewCasaOS() codegen.ServerInterface {
-	return &CasaOS{}
+	return &CasaOS{
+		fileUploadService: service.NewFileUploadService(),
+	}
 }
