@@ -39,20 +39,25 @@ func (s *connectionsStruct) GetConnectionByHost(host string) (connections []mode
 	s.db.Select("username,host,status,id").Where("host = ?", host).Find(&connections)
 	return
 }
+
 func (s *connectionsStruct) GetConnectionByID(id string) (connections model2.ConnectionsDBModel) {
 	s.db.Select("username,password,host,status,id,directories,mount_point,port").Where("id = ?", id).First(&connections)
 	return
 }
+
 func (s *connectionsStruct) GetConnectionsList() (connections []model2.ConnectionsDBModel) {
 	s.db.Select("username,host,port,status,id,mount_point").Find(&connections)
 	return
 }
+
 func (s *connectionsStruct) CreateConnection(connection *model2.ConnectionsDBModel) {
 	s.db.Create(connection)
 }
+
 func (s *connectionsStruct) UpdateConnection(connection *model2.ConnectionsDBModel) {
 	s.db.Save(connection)
 }
+
 func (s *connectionsStruct) DeleteConnection(id string) {
 	s.db.Where("id= ?", id).Delete(&model.ConnectionsDBModel{})
 }
@@ -66,9 +71,10 @@ func (s *connectionsStruct) MountSmaba(username, host, directory, port, mountPoi
 		fmt.Sprintf("username=%s,password=%s", username, password),
 	)
 	return err
-	//str := command2.ExecResultStr("source " + config.AppInfo.ShellPath + "/helper.sh ;MountCIFS " + username + " " + host + " " + directory + " " + port + " " + mountPoint + " " + password)
-	//return str
+	// str := command2.ExecResultStr("source " + config.AppInfo.ShellPath + "/helper.sh ;MountCIFS " + username + " " + host + " " + directory + " " + port + " " + mountPoint + " " + password)
+	// return str
 }
+
 func (s *connectionsStruct) UnmountSmaba(mountPoint string) error {
 	return mount.Unmount(mountPoint)
 }
